# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CleanEnergyHour {
  /**
     * @minimum 0
     * @maximum 23
     */
  hour: number;
  label: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  clean_pct: number;
}

export interface CleanEnergyToday {
  date: string;
  source_note: string;
  current_hour: number;
  curve: CleanEnergyHour[];
}

export interface ScheduleRecommendation {
  date: string;
  start_hour: number;
  end_hour: number;
  start_label: string;
  end_label: string;
  duration_hours: number;
  reason: string;
  kwh_saved: number;
  co2_avoided_kg: number;
  clean_threshold: number;
  soil_moisture: number;
  watering_threshold: number;
}

export interface ForecastResult {
  date: string;
  predicted_curve: number[];
  actual_curve: number[];
  mape: number;
  model_used: string;
  holdout_days: number;
}

export interface RipenessInput {
  /** Data URL or image identifier from the demo upload */
  image_data: string;
  image_name?: string;
}

export type RipenessResultClassification = typeof RipenessResultClassification[keyof typeof RipenessResultClassification];


export const RipenessResultClassification = {
  unripe: 'unripe',
  ripe: 'ripe',
  spoiling: 'spoiling',
} as const;

export type RipenessResultCoolingState = typeof RipenessResultCoolingState[keyof typeof RipenessResultCoolingState];


export const RipenessResultCoolingState = {
  low: 'low',
  full: 'full',
} as const;

export type RipenessResultColorMetrics = {
  green_score: number;
  ripe_score: number;
  dark_score: number;
};

export interface RipenessResult {
  timestamp: string;
  classification: RipenessResultClassification;
  confidence_note: string;
  cooling_state: RipenessResultCoolingState;
  color_metrics: RipenessResultColorMetrics;
}

export type GetScheduleParams = {
/**
 * @minimum 0
 * @maximum 100
 */
soil_moisture: number;
/**
 * @minimum 0
 * @maximum 100
 */
crop_threshold?: number;
};