# lib/api-zod/src/generated/types/cleanEnergyHour.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CleanEnergyHour {
  /**
     * @minimum 0
     * @maximum 23
     */
  hour: number;
  label: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  clean_pct: number;
}