# lib/api-zod/src/generated/types/cleanEnergyToday.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CleanEnergyHour } from './cleanEnergyHour';

export interface CleanEnergyToday {
  date: string;
  source_note: string;
  current_hour: number;
  curve: CleanEnergyHour[];
}