# lib/api-zod/src/generated/types/forecastResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ForecastResult {
  date: string;
  predicted_curve: number[];
  actual_curve: number[];
  mape: number;
  model_used: string;
  holdout_days: number;
}