# lib/api-zod/src/generated/types/ripenessInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RipenessInput {
  /** Data URL or image identifier from the demo upload */
  image_data: string;
  image_name?: string;
}