# lib/api-zod/src/generated/types/ripenessResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RipenessResultClassification } from './ripenessResultClassification';
import type { RipenessResultColorMetrics } from './ripenessResultColorMetrics';
import type { RipenessResultCoolingState } from './ripenessResultCoolingState';

export interface RipenessResult {
  timestamp: string;
  classification: RipenessResultClassification;
  confidence_note: string;
  cooling_state: RipenessResultCoolingState;
  color_metrics: RipenessResultColorMetrics;
}