# lib/api-zod/src/generated/types/scheduleRecommendation.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ScheduleRecommendation {
  date: string;
  start_hour: number;
  end_hour: number;
  start_label: string;
  end_label: string;
  duration_hours: number;
  reason: string;
  kwh_saved: number;
  co2_avoided_kg: number;
  clean_threshold: number;
  soil_moisture: number;
  watering_threshold: number;
}